fix(migration): 批次192 修复 E2E 迁移 BIGINT 外键类型不匹配
问题：m0044 整合迁移执行 31 个 SQL 文件时，多个文件的 BIGINT 外键引用了
SERIAL (INTEGER) 类型的 id 列，导致 PostgreSQL 报错。

修复内容：
1. color_card_items SQL: dye_recipes(复数) → dye_recipe(单数)，BIGINT → INTEGER
2. m0044: 新增 fix_fk_types() 函数，执行 SQL 前自动将引用 INTEGER 表的
   BIGINT 外键替换为 INTEGER 外键
3. 涉及的 INTEGER id 表：users/products/customers/dye_recipe/sales_orders/
   product_categories/warehouses/suppliers/product_colors/fixed_assets
4. 同步更新 Rust 模型：color_card_item.dye_recipe_id 从 Option<i64> → Option<i32>
5. 同步更新 DTO 和 Response DTO 的 dye_recipe_id 类型
6. 简化 color_card_scan_service 中不再需要的 i32::try_from 转换

+++ backend/src/models/color_card_item.rs
@@ -29,7 +29,7 @@ pub struct Model {
     pub cncs_code: Option<String>,
     pub custom_code: Option<String>,
     pub hex_value: String,
-    pub dye_recipe_id: Option<i64>,
+    pub dye_recipe_id: Option<i32>,
     pub product_color_price_id: Option<i64>,
     pub swatch_image_url: Option<String>,
     pub sequence: i32,
